Proguard: Fix proguard-optimized JAR not terminating after `exit`
[Why]
Enabling optimizations will turn on class merging in Proguard, which
causes Action.Undo and Action.Exit to be merged into one, leading to
Action.Exit branch in AbstractSokobanGame.processAction to never be
executed (or pruned by a subsequent optimization pass).

[How]
Disable `class/merging/horizontal` optimization, which merges
horizontal classes in the class hierarchy. This leaves Action.Undo and
Action.Exit as two distinct classes, thus preventing the issue. Other
optimizations can still be enabled as they are not shown to affect the
execution of the obfuscated JAR.